requirements.txt
torch
torchvision
transformers
datasets
peft
accelerate
pillow